+++ src/snowflake/snowpark/modin/plugin/compiler/snowflake_query_compiler.py
@@ -12,6 +12,7 @@ import typing
 import uuid
 from collections.abc import Hashable, Iterable, Mapping, Sequence
 from datetime import timedelta, tzinfo
+from functools import reduce
 from typing import Any, Callable, List, Literal, Optional, Union, get_args
 
 import modin.pandas as pd
@@ -2336,7 +2337,7 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
     def reindex(
         self,
         axis: int,
-        labels: Union[pandas.Index, "pd.Index", list[Any]],
+        labels: Union[pandas.Index, "pd.Index", list[Any], "SnowflakeQueryCompiler"],
         **kwargs: dict[str, Any],
     ) -> "SnowflakeQueryCompiler":
         """
@@ -2346,7 +2347,7 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         ----------
         axis : {0, 1}
             Axis to align labels along. 0 is for index, 1 is for columns.
-        labels : list-like
+        labels : list-like, SnowflakeQueryCompiler
             Index-labels to align with.
         method : {None, "backfill"/"bfill", "pad"/"ffill", "nearest"}
             Method to use for filling holes in reindexed frame.
@@ -2544,7 +2545,7 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
 
     def _reindex_axis_0(
         self,
-        labels: Union[pandas.Index, "pd.Index", list[Any]],
+        labels: Union[pandas.Index, "pd.Index", list[Any], "SnowflakeQueryCompiler"],
         **kwargs: dict[str, Any],
     ) -> "SnowflakeQueryCompiler":
         """
@@ -2552,7 +2553,7 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
 
         Parameters
         ----------
-        labels : list-like
+        labels : list-like, SnowflakeQueryCompiler
             Index-labels to align with.
         method : {None, "backfill"/"bfill", "pad"/"ffill", "nearest"}
             Method to use for filling holes in reindexed frame.
@@ -2570,12 +2571,15 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         """
         self._raise_not_implemented_error_for_timedelta()
 
-        if isinstance(labels, native_pd.Index):
-            labels = pd.Index(labels)
-        if isinstance(labels, pd.Index):
-            new_index_qc = labels.to_series()._query_compiler
+        if isinstance(labels, SnowflakeQueryCompiler):
+            new_index_qc = labels
         else:
-            new_index_qc = pd.Series(labels)._query_compiler
+            if isinstance(labels, native_pd.Index):
+                labels = pd.Index(labels)
+            if isinstance(labels, pd.Index):
+                new_index_qc = labels.to_series()._query_compiler
+            else:
+                new_index_qc = pd.Series(labels)._query_compiler
 
         new_index_modin_frame = new_index_qc._modin_frame
         modin_frame = self._modin_frame
@@ -4172,7 +4176,7 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         )
 
     def _fill_null_values_in_groupby(
-        self, method: str, by_list: list[str]
+        self, method: str, by_list: list[str], limit: Optional[int] = None
     ) -> dict[str, ColumnOrName]:
         """
         Fill null values in each column using method within each group.
@@ -4182,6 +4186,8 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
                 The method to use to fill null values.
             by_list: list[str]
                 The list of columns to partition by during the fillna.
+            limit : int, optional
+                The limit of values in a run to fill.
 
         Returns:
             dict: A mapping between column name and the Snowpark Column object with
@@ -4191,12 +4197,12 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         method_is_ffill = method is FillNAMethod.FFILL_METHOD
         if method_is_ffill:
             func = last_value
-            window_start = Window.UNBOUNDED_PRECEDING
+            window_start = Window.UNBOUNDED_PRECEDING if limit is None else -1 * limit
             window_end = Window.CURRENT_ROW
         else:
             func = first_value
             window_start = Window.CURRENT_ROW
-            window_end = Window.UNBOUNDED_FOLLOWING
+            window_end = Window.UNBOUNDED_FOLLOWING if limit is None else limit
 
         return {
             snowflake_quoted_id: coalesce(
@@ -5399,6 +5405,284 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
             ignore_index=not as_index,  # When as_index=False, take the default positional index
         )
 
+    def groupby_fillna(
+        self,
+        by: Any,
+        axis: int,
+        groupby_kwargs: dict[str, Any],
+        value: Optional[
+            Union[Scalar, Hashable, Mapping, "pd.DataFrame", "pd.Series"]
+        ] = None,
+        method: Optional[FillnaOptions] = None,
+        fill_axis: Optional[int] = None,
+        inplace: bool = False,
+        limit: Optional[int] = None,
+        downcast: Optional[dict] = None,
+    ) -> "SnowflakeQueryCompiler":
+        """
+        Replace NaN values using provided method or value.
+
+        Args:
+            by: Used to determine the groups for the groupby.
+            axis: Group by axis, 0 (index) or 1 (columns), only axis=0 is supported currently.
+            groupby_kwargs: Dict[str, Any]
+                keyword arguments passed for the groupby.
+            value: Optional fill value.
+            method: Optional (if no value specified) method of `ffill` or `bfill`.
+            fill_axis : Fill axis, 0 (index) or 1 (columns)
+            inplace: Not supported
+            limit: Maximum number of consecutive NA values to fill.
+            downcast: Not supported
+
+        Returns:
+            SnowflakeQueryCompiler: with a NaN values using method or value.
+        """
+        level = groupby_kwargs.get("level", None)
+
+        is_supported = check_is_groupby_supported_by_snowflake(
+            by=by, level=level, axis=axis
+        )
+        if not is_supported:
+            ErrorMessage.not_implemented(
+                f"Snowpark pandas GroupBy.fillna {_GROUPBY_UNSUPPORTED_GROUPING_MESSAGE}"
+            )
+
+        if by is not None and not is_list_like(by):
+            by = [by]
+
+        if value is not None and method is not None:
+            raise ValueError("Cannot specify both 'value' and 'method'.")
+
+        if value is None and method is None:
+            raise ValueError("Must specify a fill 'value' or 'method'.")
+
+        if method is not None and method not in ["ffill", "bfill"]:
+            raise ValueError(
+                f"Invalid fill method. Expecting pad (ffill) or backfill (bfill). Got {method}"
+            )
+
+        if downcast:
+            ErrorMessage.not_implemented(
+                "Snowpark pandas fillna API doesn't yet support 'downcast' parameter"
+            )
+
+        if fill_axis is None:
+            fill_axis = 0
+
+        if level is not None:
+            by = extract_groupby_column_pandas_labels(self, by, level)
+
+        frame = self._modin_frame
+
+        data_column_group_keys = [
+            pandas_label
+            for pandas_label in frame.data_column_pandas_labels
+            if pandas_label in by
+        ]
+
+        data_column_group_keys_mask = [
+            pandas_label in data_column_group_keys
+            for pandas_label in frame.data_column_pandas_labels
+        ]
+
+        by_list_snowflake_quoted_identifiers: list[str]
+
+        # If any of the groupby values are None, then there is no fill so check through an expression.
+        def groupby_null_expr(
+            col_expr: SnowparkColumn,
+            col_snowflake_quoted_identifier: str,
+            by_list_snowflake_quoted_identifiers: List[str],
+        ) -> SnowparkColumn:
+            return iff(
+                reduce(
+                    lambda b1, b2: b1 & b2,
+                    [
+                        (col(snowflake_quoted_identifier).is_not_null())
+                        for snowflake_quoted_identifier in by_list_snowflake_quoted_identifiers
+                    ],
+                ),
+                col_expr,
+                pandas_lit(None),
+            ).as_(col_snowflake_quoted_identifier)
+
+        # If no method, then we will use the value instead.
+        if method is None:
+            # If there's no method, then the fill is same as dataframe.fillna with fill value.  Skip any group by
+            # data columns in the fill.
+            qc = self._fillna_with_masking(
+                value=value,
+                self_is_series=False,
+                method=None,
+                axis=axis,
+                limit=limit,
+                downcast=downcast,
+                columns_mask=data_column_group_keys_mask,
+            )
+
+            frame = qc._modin_frame
+            new_snowflake_quoted_identifiers = (
+                frame.data_column_snowflake_quoted_identifiers
+            )
+
+            # Group by snowflake quoted identifiers
+            by_list_snowflake_quoted_identifiers = [
+                snowflake_quoted_identifier[0]
+                for snowflake_quoted_identifier in frame.get_snowflake_quoted_identifiers_group_by_pandas_labels(
+                    by, include_index=True, include_data=True
+                )
+                if len(snowflake_quoted_identifier) > 0
+            ]
+
+            # Generate new snowflake quoted identifiers for output columns so they don't conflict with existing.
+            new_snowflake_quoted_identifiers = (
+                frame.ordered_dataframe.generate_snowflake_quoted_identifiers(
+                    pandas_labels=frame.data_column_pandas_labels
+                )
+            )
+
+            # Select the resulting columns excluding group by values which are None.
+            select_list = frame.index_column_snowflake_quoted_identifiers + [
+                groupby_null_expr(
+                    col(snowflake_quoted_identifier),
+                    new_snowflake_quoted_identifier,
+                    by_list_snowflake_quoted_identifiers,
+                )
+                for new_snowflake_quoted_identifier, snowflake_quoted_identifier in zip(
+                    new_snowflake_quoted_identifiers,
+                    frame.data_column_snowflake_quoted_identifiers,
+                )
+            ]
+        else:
+            # Group by snowflake quoted identifiers
+            by_list_snowflake_quoted_identifiers = [
+                snowflake_quoted_identifier[0]
+                for snowflake_quoted_identifier in frame.get_snowflake_quoted_identifiers_group_by_pandas_labels(
+                    by, include_index=True, include_data=True
+                )
+                if len(snowflake_quoted_identifier) > 0
+            ]
+
+            # Generate new snowflake quoted identifiers for output columns so they don't conflict with existing.
+            new_snowflake_quoted_identifiers = (
+                frame.ordered_dataframe.generate_snowflake_quoted_identifiers(
+                    pandas_labels=frame.data_column_pandas_labels
+                )
+            )
+
+            if fill_axis == 0:
+                # Fill the groups row-wise with values.
+                columns_to_fillna_expr = self._fill_null_values_in_groupby(
+                    method, by_list_snowflake_quoted_identifiers, limit
+                )
+
+                select_list = frame.index_column_snowflake_quoted_identifiers + [
+                    groupby_null_expr(
+                        columns_to_fillna_expr[snowflake_quoted_identifier],
+                        new_snowflake_quoted_identifier,
+                        by_list_snowflake_quoted_identifiers,
+                    )
+                    for new_snowflake_quoted_identifier, snowflake_quoted_identifier in zip(
+                        new_snowflake_quoted_identifiers,
+                        frame.data_column_snowflake_quoted_identifiers,
+                    )
+                ]
+
+            elif fill_axis == 1:
+                # Fill the groups column-wise using coalesce of prior columns depending on method direction.
+                coalesce_column_list: list[SnowparkColumn] = []
+                select_list = []
+
+                data_column_pairs = list(
+                    zip(
+                        frame.data_column_snowflake_quoted_identifiers,
+                        new_snowflake_quoted_identifiers,
+                        data_column_group_keys_mask,
+                    )
+                )
+
+                # If we are doing a ffill, then the current NaN value will fill in with the values from the left,
+                # otherwise if bfill will fill with values from the right.  To simplify the processing, we reverse
+                # the column order for bfill so it's the same pass to generate the coalece of prior columns, and
+                # then reverse again afterwards.
+                if method == "bfill":
+                    data_column_pairs.reverse()
+
+                for (
+                    snowflake_quoted_identifier,
+                    new_snowflake_quoted_identifier,
+                    is_data_column_group_key,
+                ) in data_column_pairs:
+                    if is_data_column_group_key:
+                        select_list.append(
+                            col(snowflake_quoted_identifier).as_(
+                                new_snowflake_quoted_identifier
+                            )
+                        )
+                        continue
+
+                    if len(coalesce_column_list) == 0:
+                        select_item = col(snowflake_quoted_identifier)
+                    else:
+                        coalesce_expr = [
+                            snowflake_quoted_identifier
+                        ] + coalesce_column_list
+                        select_item = coalesce(*coalesce_expr)
+
+                    select_item = groupby_null_expr(
+                        select_item,
+                        new_snowflake_quoted_identifier,
+                        by_list_snowflake_quoted_identifiers,
+                    )
+
+                    select_list.append(select_item)
+                    coalesce_column_list.insert(0, col(snowflake_quoted_identifier))
+                    if limit is not None and len(coalesce_column_list) > limit:
+                        del coalesce_column_list[-1]
+
+                if method == "bfill":
+                    select_list.reverse()
+
+                select_list = (
+                    frame.index_column_snowflake_quoted_identifiers + select_list
+                )
+
+        new_ordered_dataframe = frame.ordered_dataframe.select(select_list)
+
+        # If any group-by keys were original data (not index) columns, then we drop them in the final result.
+        #
+        #     The methods ffill, bfill, pad and backfill of DataFrameGroupBy previously included the group labels in
+        #      the return value, which was inconsistent with other groupby transforms. Now only the filled values
+        #      are returned. (GH 21521)
+        if len(data_column_group_keys) > 0:
+            data_column_pandas_labels, data_column_snowflake_quoted_identifiers = zip(
+                *[
+                    (pandas_label, snowflake_quoted_identifier)
+                    for pandas_label, snowflake_quoted_identifier in zip(
+                        frame.data_column_pandas_labels,
+                        new_snowflake_quoted_identifiers,
+                    )
+                    if pandas_label not in data_column_group_keys
+                ]
+            )
+        else:
+            data_column_pandas_labels, data_column_snowflake_quoted_identifiers = (
+                frame.data_column_pandas_labels,
+                new_snowflake_quoted_identifiers,
+            )
+
+        new_frame = InternalFrame.create(
+            ordered_dataframe=new_ordered_dataframe,
+            index_column_pandas_labels=frame.index_column_pandas_labels,
+            index_column_snowflake_quoted_identifiers=frame.index_column_snowflake_quoted_identifiers,
+            data_column_pandas_labels=data_column_pandas_labels,
+            data_column_snowflake_quoted_identifiers=data_column_snowflake_quoted_identifiers,
+            data_column_pandas_index_names=frame.data_column_pandas_index_names,
+            data_column_types=None,
+            index_column_types=None,
+        )
+
+        return SnowflakeQueryCompiler(new_frame)
+
     def _get_dummies_helper(
         self,
         column: Hashable,
@@ -6234,7 +6518,7 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         # 'loc'
         def move_last_element(arr: list, index: int) -> None:
             if replace:
-                # swap element at loc with new colun at end, then drop last element
+                # swap element at loc with new column at end, then drop last element
                 arr[index], arr[-1] = arr[-1], arr[index]
                 arr.pop()
             else: