+++ src/snowflake/snowpark/modin/plugin/docstrings/groupby.py
@@ -1885,7 +1885,94 @@ class DataFrameGroupBy:
         pass
 
     def fillna():
-        pass
+        """
+        Fill NA/NaN values using the specified method within groups.
+
+        Parameters
+        ----------
+        value : scalar, dict, Series, or DataFrame
+            value to use to fill holes (e.g. 0), alternately a dict/Series/DataFrame of values
+            specifying which value to use for each index (for a Series) or column (for a
+            DataFrame). Values not in the dict/Series/DataFrame will not be filled. This
+            value cannot be a list.
+
+        method : {{‘bfill’, ‘ffill’, None}}, default None
+            Method to use for filling holes. 'ffill' will propagate the last valid observation
+            forward within a group. 'bfill' will use next valid observation to fill the gap.
+
+        axis : {0 or ‘index’, 1 or ‘columns’}
+            Axis along which to fill missing values. When the DataFrameGroupBy axis
+            argument is 0, using axis=1 here will produce the same results as
+            DataFrame.fillna(). When the DataFrameGroupBy axis argument is 1, using
+            axis=0 or axis=1 here will produce the same results.
+
+        inplace : bool, default False
+            Ignored.
+
+        limit : int, default None
+            If method is specified, this is the maximum number of consecutive NaN values to
+            forward/backward fill within a group. In other words, if there is a gap with more than
+            this number of consecutive NaNs, it will only be partially filled. If method is not
+            specified, this is the maximum number of entries along the entire axis where NaNs
+            will be filled. Must be greater than 0 if not None.
+
+        downcast : dict, default is None
+            A dict of item->dtype of what to downcast if possible, or the string ‘infer’ which will
+            try to downcast to an appropriate equal type (e.g. float64 to int64 if possible).
+
+            This parameter is not yet supported in Snowpark pandas.
+
+        Returns
+        -------
+        :class:`~modin.pandas.DataFrame`
+           Object with missing values filled.
+
+        Examples
+        --------
+        >>> df = pd.DataFrame(
+        ...     {
+        ...         "key": [0, 0, 1, 1, 1],
+        ...         "A": [np.nan, 2, np.nan, 3, np.nan],
+        ...         "B": [2, 3, np.nan, np.nan, np.nan],
+        ...         "C": [np.nan, np.nan, 2, np.nan, np.nan],
+        ...     }
+        ... )
+        >>> df
+           key    A    B    C
+        0    0  NaN  2.0  NaN
+        1    0  2.0  3.0  NaN
+        2    1  NaN  NaN  2.0
+        3    1  3.0  NaN  NaN
+        4    1  NaN  NaN  NaN
+
+        Propagate non-null values forward or backward within each group along columns.
+
+        >>> df.groupby("key").fillna(method="ffill")
+             A    B    C
+        0  NaN  2.0  NaN
+        1  2.0  3.0  NaN
+        2  NaN  NaN  2.0
+        3  3.0  NaN  2.0
+        4  3.0  NaN  2.0
+
+        >>> df.groupby("key").fillna(method="bfill")
+             A    B    C
+        0  2.0  2.0  NaN
+        1  2.0  3.0  NaN
+        2  3.0  NaN  2.0
+        3  3.0  NaN  NaN
+        4  NaN  NaN  NaN
+
+        Only replace the first NaN element within a group along rows.
+
+        >>> df.groupby("key").fillna(method="ffill", limit=1)
+             A    B    C
+        0  NaN  2.0  NaN
+        1  2.0  3.0  NaN
+        2  NaN  NaN  2.0
+        3  3.0  NaN  2.0
+        4  3.0  NaN  NaN
+        """
 
     def count():
         """