+++ src/snowflake/snowpark/modin/plugin/extensions/series_extensions.py
@@ -17,6 +17,9 @@ from snowflake.snowpark.dataframe import DataFrame as SnowparkDataFrame
 from snowflake.snowpark.modin import pandas as pd  # noqa: F401
 from snowflake.snowpark.modin.pandas.api.extensions import register_series_accessor
 from snowflake.snowpark.modin.plugin.extensions.utils import add_cache_result_docstring
+from snowflake.snowpark.modin.plugin.utils.warning_message import (
+    materialization_warning,
+)
 
 
 @register_series_accessor("to_snowflake")
@@ -167,6 +170,7 @@ def to_snowpark(
 
 
 @register_series_accessor("to_pandas")
+@materialization_warning
 def to_pandas(
     self,
     *,
@@ -200,6 +204,7 @@ def to_pandas(
 
 @register_series_accessor("cache_result")
 @add_cache_result_docstring
+@materialization_warning
 def cache_result(self, inplace: bool = False) -> Optional[pd.Series]:
     """
     Persists the Snowpark pandas Series to a temporary table for the duration of the session.